refactor: split addScoreScriptToHtml Function
